Adiciona middleware de autenticação e atualiza plano de refatoração
