+++ docs/plano-refatoracao-backend.md
@@ -149,13 +149,12 @@ Este documento serve como guia e controle de progresso para a refatoração arqu
 ### 🛡️ Fase 4: Middlewares e Validação (Semana 7)
 
 #### ✅ Tarefas Concluídas
-- [ ] Nenhuma tarefa concluída
+- [x] Criar middleware de autenticação centralizado
 
 #### 🔄 Em Andamento
 - [ ] Nenhuma tarefa em andamento
 
 #### ⏳ Pendentes
-- [ ] Criar middleware de autenticação centralizado
 - [ ] Implementar middleware de validação com Zod
 - [ ] Criar middleware de tratamento de erros
 - [ ] Implementar middleware de logging
@@ -202,9 +201,9 @@ Este documento serve como guia e controle de progresso para a refatoração arqu
 ## 📊 Métricas de Progresso
 
 ### 📈 Status Geral
-  - **Progresso Total**: 30% (1/6 fases concluídas)
+  - **Progresso Total**: 32% (1/6 fases concluídas)
   - **Fase Atual**: Fase 2 - Implementação de Serviços
-  - **Tarefas Concluídas**: 18/60 (30%)
+  - **Tarefas Concluídas**: 19/60 (32%)
   - **Próxima Milestone**: Implementação de Serviços
 
 ### 📋 Checklist por Fase
@@ -214,7 +213,7 @@ Este documento serve como guia e controle de progresso para a refatoração arqu
 | 1 | Reestruturação de Rotas | ✅ Concluída | 12/12 (100%) |
 | 2 | Implementação de Serviços | 🚧 Em andamento | 6/8 (75%) |
 | 3 | Padrão Repository | ⏳ Pendente | 0/8 (0%) |
-| 4 | Middlewares e Validação | ⏳ Pendente | 0/7 (0%) |
+| 4 | Middlewares e Validação | 🚧 Em andamento | 1/7 (14%) |
 | 5 | Controllers | ⏳ Pendente | 0/8 (0%) |
 | 6 | Otimizações | ⏳ Pendente | 0/8 (0%) |
