+++ server/routes/product.routes.ts
@@ -1,5 +1,5 @@
 import type { Express } from "express";
-import { isAuthenticated } from "../auth";
+import { isAuthenticated } from "../middlewares/auth.middleware";
 import { productService } from "../services/product.service";
 
 export function registerProductRoutes(app: Express) {
